# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginSessionRead": {
        "description": "A live interactive-login session.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "site_id": {
            "title": "Site Id",
            "type": "integer"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "ws_path": {
            "title": "Ws Path",
            "type": "string"
          }
        },
        "required": [
          "site_id",
          "ws_path",
          "width",
          "height",
          "expires_at"
        ],
        "title": "LoginSessionRead",
        "type": "object"
      },
      "Message": {
        "description": "A plain result for actions that have nothing else to say.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Message",
        "type": "object"
      },
      "PingMethod": {
        "description": "How a site is pinged.",
        "enum": [
          "http",
          "browser"
        ],
        "title": "PingMethod",
        "type": "string"
      },
...
      "SiteRead": {
        "description": "A site as the UI shows it.",
        "properties": {
          "consecutive_failures": {
            "title": "Consecutive Failures",
            "type": "integer"
          },
          "deadline_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "expected_status": {
            "default": 200,
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Expected Status",
            "type": "integer"
          },
          "failure_pattern": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Pattern"
          },
          "favicon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favicon"
          },
          "follow_redirects": {
            "default": true,
            "title": "Follow Redirects",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "inactivity_limit_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactivity Limit Days"
          },
          "interval_days": {
            "default": 7,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Interval Days",
            "type": "integer"
          },
          "jitter_percent": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Jitter Percent",
            "type": "integer"
          },
          "last_check_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Check At"
          },
          "last_ok_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ok At"
          },
          "login_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Login Url"
          },
          "login_url_pattern": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Login Url Pattern"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "next_check_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Check At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "ping_method": {
            "$ref": "#/components/schemas/PingMethod",
            "default": "http"
          },
          "ping_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Ping Url",
            "type": "string"
          },
          "pulse": {
            "items": {
              "$ref": "#/components/schemas/CheckOutcome"
            },
            "title": "Pulse",
            "type": "array"
          },
          "risk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          },
          "session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/SiteStatus"
          },
          "success_pattern": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Success Pattern"
          },
          "user_agent": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "name",
          "ping_url",
          "id",
          "status",
          "consecutive_failures",
          "last_check_at",
          "last_ok_at",
          "next_check_at",
          "deadline_at"
        ],
        "title": "SiteRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Keep logged-in website sessions alive with scheduled pings",
    "title": "lifeline",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sites": {
      "get": {
        "description": "Every site, with the recent history the list displays.",
        "operationId": "index_api_sites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteRead"
                  },
                  "title": "Response Index Api Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Index",
        "tags": [
          "sites"
        ]
      },
      "post": {
        "description": "Add a site.",
        "operationId": "create_api_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}": {
      "delete": {
        "description": "Remove a site, its stored session and its history.",
        "operationId": "destroy_api_sites__site_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Destroy",
        "tags": [
          "sites"
        ]
      },
      "get": {
        "description": "One site.",
        "operationId": "show_api_sites__site_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show",
        "tags": [
          "sites"
        ]
      },
      "put": {
        "description": "Change a site's configuration.",
        "operationId": "update_api_sites__site_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/check": {
      "post": {
        "description": "Ping a site immediately and return the result.",
        "operationId": "check_now_api_sites__site_id__check_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Now",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/checks": {
      "get": {
        "description": "A site's recent checks, newest first.",
        "operationId": "history_api_sites__site_id__checks_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CheckRead"
                  },
                  "title": "Response History Api Sites  Site Id  Checks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/detect": {
      "post": {
        "description": "Work out how to tell this site's live session from a dead one.\n\nFetches the ping URL twice, once with the stored session and once with no cookies, and\nreports what differs. Needs a session, because half the comparison is what the page looks\nlike to somebody who is logged in.",
        "operationId": "detect_api_sites__site_id__detect_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectedRules"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/login-session": {
      "post": {
        "description": "Start a browser at the site's login page and stream its screen.",
        "operationId": "open_login_api_sites__site_id__login_session_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginSessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open Login",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/session": {
      "delete": {
        "description": "Discard a site's stored session.",
        "operationId": "forget_session_api_sites__site_id__session_delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forget Session",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/session/import": {
      "post": {
        "description": "Store a session pasted in by hand.",
        "operationId": "import_session_api_sites__site_id__session_import_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CookieImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Session",
        "tags": [
          "sites"
        ]
      }
    },
    "/api/sites/{site_id}/test-rules": {
      "post": {
        "description": "Try a set of rules against the site as it is now, signed in and signed out.\n\nAnswers the question someone actually has \u2014 would this notice a dead session? \u2014 by running\nthe same verdict the scheduler runs, on both pages, with the rules currently on screen.",
        "operationId": "test_rules_api_sites__site_id__test_rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTrial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleTrialResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Rules",
        "tags": [
          "sites"
        ]
      }
    }
  }
}